refactor!: change factory from class to method, and separate loader.
