+++ cliboa/core/factory.py
@@ -13,32 +13,42 @@
 #
 from importlib import import_module
 
-from cliboa.core.strategy import MultiProcExecutor, SingleProcExecutor
+from cliboa.core.loader import _JsonScenarioLoader, _ScenarioLoader, _YamlScenarioLoader
+from cliboa.core.strategy import MultiProcExecutor, SingleProcExecutor, _StepExecutor
 from cliboa.scenario.base import BaseStep
 from cliboa.util.class_util import ClassUtil
+from cliboa.util.exception import InvalidFormat
 from cliboa.util.parallel_with_config import ParallelWithConfig
 
 
-class StepExecutorFactory:
+def _get_scenario_loader_class(scenario_format: str) -> type[_ScenarioLoader]:
+    """
+    Create scenario loader instance
+    """
+    if scenario_format == "yaml":
+        return _YamlScenarioLoader
+    elif scenario_format == "json":
+        return _JsonScenarioLoader
+    else:
+        raise InvalidFormat(f"scenario format '{scenario_format}' is invalid.")
+
+
+def _create_step_executor(obj: BaseStep | ParallelWithConfig) -> _StepExecutor:
     """
     Create step execution strategy instance
+
+    Args:
+        obj: queue which stores execution target steps
+    Returns:
+        step execution strategy instance
     """
-
-    @staticmethod
-    def create(obj: BaseStep | ParallelWithConfig):
-        """
-        Args:
-            obj: queue which stores execution target steps
-        Returns:
-            step execution strategy instance
-        """
-        if isinstance(obj, ParallelWithConfig):
-            return MultiProcExecutor(obj)
-        else:
-            return SingleProcExecutor(obj)
+    if isinstance(obj, ParallelWithConfig):
+        return MultiProcExecutor(obj)
+    else:
+        return SingleProcExecutor(obj)
 
 
-class CustomInstanceFactory(object):
+def _create_custom_instance(cls_name: str):
     """
     Import python module and create instance dynamically
 
@@ -46,12 +56,9 @@ class CustomInstanceFactory(object):
         Created instance.
         None: If cls_name was not found in the defined class list.
     """
-
-    @staticmethod
-    def create(cls_name):
-        ret = ClassUtil().describe_class(cls_name)
-        if ret is None:
-            return None
-        (root, mod_name) = ret
-        instance = getattr(import_module(root), mod_name)
-        return instance()
+    ret = ClassUtil().describe_class(cls_name)
+    if ret is None:
+        return None
+    (root, mod_name) = ret
+    instance = getattr(import_module(root), mod_name)
+    return instance()
